PCHR-3205: Replace TaskType fabricator with OptionValue fabricator
